+++ sht3x/sht3x_example_usage.c
@@ -45,7 +45,7 @@ int main(void) {
     /* Busy loop for initialization, because the main loop does not work without
      * a sensor.
      */
-    while (sht3x_probe() != STATUS_OK) {
+    while (sht3x_probe(SHT3X_I2C_ADDR_DFLT) != STATUS_OK) {
         printf("SHT sensor probing failed\n");
     }
     printf("SHT sensor probing successful\n");
@@ -55,7 +55,8 @@ int main(void) {
         /* Measure temperature and relative humidity and store into variables
          * temperature, humidity (each output multiplied by 1000).
          */
-        int8_t ret = sht3x_measure_blocking_read(&temperature, &humidity);
+        int8_t ret = sht3x_measure_blocking_read(SHT3X_I2C_ADDR_DFLT,
+                                                 &temperature, &humidity);
         if (ret == STATUS_OK) {
             printf("measured temperature: %0.2f degreeCelsius, "
                    "measured humidity: %0.2f percentRH\n",