+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+ * [`changed`] SHT3x driver added macros to easily parse the STATUS register relevant bits
+ * [`changed`] SHT3x driver support for alert thresholds configuration (read and write commands)
+ * [`changed`] SHT3x driver support for clear status register
+ * [`changed`] SHT3x driver allows use of 2 sensors in parallel (addresses 0x44 & 0x45)
+
 ## [5.2.1] - 2020-12-14
 
  * [`changed`] Makefile to only include needed files from embedded-common